multiagent-test/requirements.txt
# Core ADK (with A2A support)
google-adk[a2a]
google-genai

# AG-UI Integration (for calendar agent)
ag-ui-adk

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP Client (for A2A communication)
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# CORS support
python-multipart>=0.0.6
